requirements.txt
# =============================================================================
# Wolin Learn - Python 依赖包
# =============================================================================
# AI/LLM 相关
openai>=1.90.0
ollama>=0.6.1
dashscope>=1.20.0

# LangChain & LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
langchain-ollama>=0.1.0
langgraph>=0.2.0

# 向量数据库
pymilvus>=2.4.0

# Embedding 模型
sentence-transformers>=3.0.0

# Web 框架
fastapi>=0.109.0
uvicorn>=0.27.0
sse-starlette>=2.0.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.12.0
httpx>=0.27.0
